cleanup: refactor tests, remove unused imports
