+++ .gitignore
@@ -1,3 +1,9 @@
 __pycache__
 GraFT/still
 GraFT/timeseries
+build
+dist
+*.egg-info
+venv*/
+graft/output/
+graft/still/
